[Move] 알람 관련 쿼리문 alarm resolver로 이동
friendAlarm에 알람 관련 쿼리문들을 모아 비대해진 friend를 줄이려고 하였다.
